requirements.txt
streamlit==1.48.0
python-docx==1.2.0
PyPDF2==3.0.1
openai==1.99.6
Pillow==11.3.0
docx2txt==0.9